projs_test/requirements.txt
certifi==2017.11.5
chardet==3.0.4
class-registry==2.1.0
click==6.7
coreapi==2.3.3
coreapi-cli==1.0.6
coreschema==0.0.4
decorator==4.1.2
Django==1.11.7
django-filter==1.1.0
djangorestframework==3.7.3
filters==1.3.2
idna==2.6
ipython==6.2.1
ipython-genutils==0.2.0
itypes==1.1.0
jedi==0.11.0
Jinja2==2.10
Markdown==2.6.9
MarkupSafe==1.0
parso==0.1.0
pexpect==4.3.0
pickleshare==0.7.4
prompt-toolkit==1.0.15
ptyprocess==0.5.2
Pygments==2.2.0
PyMySQL==0.7.11
python-dateutil==2.6.1
pytz==2017.3
regex==2017.12.5
requests==2.18.4
simplegeneric==0.8.1
six==1.11.0
traitlets==4.3.2
uritemplate==3.0.0
urllib3==1.22
wcwidth==0.1.7
